+++ src/main/java/com/github/black0nion/blackonionbot/bot/SlashCommandBase.java
@@ -1,13 +1,13 @@
 package com.github.black0nion.blackonionbot.bot;
 
-import com.github.black0nion.blackonionbot.commands.SlashCommand;
-import com.github.black0nion.blackonionbot.commands.SlashCommandEvent;
-import com.github.black0nion.blackonionbot.commands.admin.BanUsageCommand;
-import com.github.black0nion.blackonionbot.commands.bot.HelpCommand;
-import com.github.black0nion.blackonionbot.commands.bot.ToggleCommand;
+import com.github.black0nion.blackonionbot.slashcommands.SlashCommand;
+import com.github.black0nion.blackonionbot.slashcommands.SlashCommandEvent;
+import com.github.black0nion.blackonionbot.slashcommands.admin.BanUsageCommand;
+import com.github.black0nion.blackonionbot.slashcommands.bot.ToggleCommand;
 import com.github.black0nion.blackonionbot.misc.Category;
 import com.github.black0nion.blackonionbot.misc.GuildType;
 import com.github.black0nion.blackonionbot.misc.RunMode;
+import com.github.black0nion.blackonionbot.slashcommands.information.HelpCommand;
 import com.github.black0nion.blackonionbot.stats.StatisticsManager;
 import com.github.black0nion.blackonionbot.systems.dashboard.Dashboard;
 import com.github.black0nion.blackonionbot.utils.*;
