+++ src/main/java/com/github/black0nion/blackonionbot/slashcommands/bot/SupportCommand.java
@@ -17,7 +17,7 @@ import static com.github.black0nion.blackonionbot.utils.config.Config.metadata;
 public class SupportCommand extends SlashCommand {
 
     public SupportCommand() {
-        super(builder(Commands.slash("support", "Used to get support.")));
+        super("support", "Used to get support.");
     }
 
     private static final String bot_developers = metadata.authors().entrySet().stream().map(entry -> "[" + entry.getKey() + "](" + entry.getValue() + ")").collect(Collectors.joining(", "));
@@ -26,8 +26,8 @@ public class SupportCommand extends SlashCommand {
     @Override
     public void execute(@NotNull SlashCommandEvent cmde, SlashCommandInteractionEvent e, BlackMember member, BlackUser author, BlackGuild guild, TextChannel channel) {
         cmde.reply(cmde.success()
-                .addField("Get help", "[Discord Server](https://dsc.gg/blackonion)", false)
-                .addField("Bot Developers", bot_developers, false)
-                .addField("BlackOnion Authors", blackonion_authors, false));
+                .addField("gethelp", "[Discord Server](https://dsc.gg/blackonion)", false)
+                .addField("botdevelopers", bot_developers, false)
+                .addField("blackoniondevs", blackonion_authors, false));
     }
 }
